ddl, store: move the ddl dependency from tikv pkg.
